There we go
No need to panick. Except dicks are gone. Sorry about that. However, we
still need a title sprite, a start sprite, a continue sprite, a load
sprite, an options sprite, and a quit sprite for the menu. Also maybe a
background for the future. They should be fairly small, but not too
small that you can't click them.