Added Some stuff
I pseudo fixed the menu_arrow to align with the lettering. I also added
a player logic which also stops moving when global.pause = 1. I also
created an options menu but it doesn't do anything right now. I also
edited the scr_menu to refer to the index of the rooms rather than the
next room.